Refactor arg codegen
Now a function is generated for each field and directive that has
arguments. This function can be used by both field methods as well as
the `Complexity` method.

The `args.gotpl` template now generates the code for this function, so
its purpose is a little different than it used to be.