Add Link to Sample Project with GQLGen and Postgres
This is a very straightforward project with numerous details in the README and the official
documentation, but questions continue to pop up around how to use this project, organize the files
and ultimately make data calls to some persistent layer.

The `https://github.com/oshalygin/gqlgen-pg-todo-example` was built in order to show newcomers the
following:
- How to organize their graphql schema, resolvers, dataloaders and models
- How to create a new dataloader
- How to resolve with a dataloader and how to avoid some of the pitfalls(inconsistent db query to keys array order)
- How to map models from a gql schema to structs

While the examples in this project are helpful, they could benefit from more elaborate explanations in the
code as well as the README to help newcomers get started.  This PR is not intended to portray any of the examples
negatively and should not be interpreted as such.  There are many findings/lessons learned from the work that folks
put together in those examples.

README which covers a ton of the details on how to use this project:
- [README](https://github.com/oshalygin/gqlgen-pg-todo-example)